hello-world-agent/pyproject.toml
[project]
name = "hello-world-agent"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "IBM Corp." }]
requires-python = ">=3.11,<4.0"
dependencies = [
    "acp-sdk", # version determined by beeai-sdk
    "beeai-sdk==0.0.21",
]

[project.scripts]
server = "hello_world_agent.server:server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
